Move MqttMapping to connector
